+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add option for spawned buildings to respond to underlying zoning changes.
+
 ### 0.7.1
 - Compatibility with game version 1.1.6.